@@ -1,55 +1,44 @@
-import React, { Component } from 'react';
-import { titleCase } from '../utilities.js';
-import axios from 'axios';
-import Select from 'react-select';
-import 'react-select/dist/react-select.css';
-import '../styles/favorite_team_select.css';
+import React, { Component } from 'react'
+import { generateTeamOptions } from '../utilities.js'
+import axios from 'axios'
+import Select from 'react-select'
+import 'react-select/dist/react-select.css'
+import '../styles/favorite_team_select.css'
 
 class FavoriteTeamSelect extends Component {
-  constructor() {
-    super();
-    this.state = {
+  state = {
       teams: []
-    };
   }
 
-  generateTeamOptions = () => {
-    return this.state.teams.map(function(team) {
-      let tmp = {}
-      tmp['value'] = team.slug;
-      tmp['label'] = titleCase(team.name);
-      tmp['className'] = 'favorite-team-option'
-      return tmp;
-    });
-  }
-
-  componentWillMount(){
+  componentDidMount(){
     axios.get('/teams')
     .then(response => {
       this.setState({
         teams: response.data || []
-      });
+      })
     })
   }
 
   render(){
-    const { teams } = this.state;
-    const { favoriteTeamSlugs, handleFavoriteTeamChange } = this.props;
-    const teamOptions = this.generateTeamOptions();
+    const { teams } = this.state
+    const { favoriteTeams, handleFavoriteTeamChange } = this.props
+    const teamOptions = generateTeamOptions(teams)
+    const favoriteTeamSlugs = favoriteTeams.map((team) => team['slug'])
+
     return(
       <div id="favorite-team-select-container flex-grow">
-        <label className="label">Favorite Teams</label>
         <Select
+          className="favorite-team-select-multiple"
           name="favorite-team-multiselect"
           value={ favoriteTeamSlugs }
           onChange={ handleFavoriteTeamChange }
           options={ teamOptions }
           multi={true}
-          closeOnSelect={false}
+          closeOnSelect={true}
         />
       </div>
-    );
+    )
   }
 }
 
-export default FavoriteTeamSelect;
+export default FavoriteTeamSelect
