@@ -1,21 +1,29 @@
 import React, { Component } from 'react';
+import { Link } from 'react-router-dom';
 import logo from '../images/sportcasts-logo.png';
-import '../styles/dashboard.css';
+import '../styles/navbar.css';
+import NavbarUserItem from './NavbarUserItem';
 
 class Navbar extends Component {
   render() {
     return (
       <nav className="navbar" aria-label="main navigation">
         <div className="navbar-brand">
-          <a className="navbar-item" href="/">
+          <Link to="/" className="navbar-item">
             <img id="logo" src={logo} alt="Logo" />
-          </a>
+          </Link>
 
           <button className="button navbar-burger">
             <span></span>
             <span></span>
             <span></span>
           </button>
+        </div>
+
+        <div id="navbarExampleTransparentExample" className="navbar-menu">
+          <div className="navbar-end">
+            <NavbarUserItem />
+          </div>
         </div>
       </nav>
     );
