@@ -2,65 +2,29 @@ Rails.application.routes.draw do
   # The priority is based upon order of creation: first created -> highest priority.
   # See how all your routes lay out with "rake routes".
 
-  get '/scrape/:league', to: 'games#scrape'
-  post '/games', to: 'games#retrieve_given_params'
-  post '/zip_codes', to: 'user_zip_codes#update'
-  post '/retrieve_zip', to: 'user_zip_codes#show'
+  resources :users do
+    resources :phones, controller: 'user_phones' do
+      post 'verify', on: :member
+      post 'resend_verification_code', on: :member
+    end
+  end
+  get  '/scrape/:league',        to: 'games#scrape'
+  post '/games',                 to: 'games#retrieve_given_params'
+  post '/zip_codes',             to: 'user_zip_codes#update'
+  post '/retrieve_zip',          to: 'user_zip_codes#show'
+  post '/update_favorite_teams', to: 'user_favorite_teams#update_all'
 
   resources :games
-
+  resources :users, only: [:create]
   resources :teams do
     resources :nicknames
   end
-  # You can have the root of your site routed with "root"
-  # root 'welcome#index'
+  resources :user_notification_preferences, only: [:index, :create, :destroy, :update]
+  resources  :user_favorite_teams
 
-  # Example of regular route:
-  #   get 'products/:id' => 'catalog#view'
 
-  # Example of named route that can be invoked with purchase_url(id: product.id)
-  #   get 'products/:id/purchase' => 'catalog#purchase', as: :purchase
-
-  # Example resource route (maps HTTP verbs to controller actions automatically):
-  #   resources :products
-
-  # Example resource route with options:
-  #   resources :products do
-  #     member do
-  #       get 'short'
-  #       post 'toggle'
-  #     end
-  #
-  #     collection do
-  #       get 'sold'
-  #     end
-  #   end
-
-  # Example resource route with sub-resources:
-  #   resources :products do
-  #     resources :comments, :sales
-  #     resource :seller
-  #   end
-
-  # Example resource route with more complex sub-resources:
-  #   resources :products do
-  #     resources :comments
-  #     resources :sales do
-  #       get 'recent', on: :collection
-  #     end
-  #   end
-
-  # Example resource route with concerns:
-  #   concern :toggleable do
-  #     post 'toggle'
-  #   end
-  #   resources :posts, concerns: :toggleable
-  #   resources :photos, concerns: :toggleable
-
-  # Example resource route within a namespace:
-  #   namespace :admin do
-  #     # Directs /admin/products/* to Admin::ProductsController
-  #     # (app/controllers/admin/products_controller.rb)
-  #     resources :products
-  #   end
+  # Sessions
+  get    '/login',  to: 'sessions#new'
+  post   '/login',  to: 'sessions#create'
+  delete '/logout', to: 'sessions#destroy'
 end
